토스트 에디터 이미지 url 축소, genfile로 저장,post 삭제 시 이미지도 삭제

+++ src/main/java/com/ll/medium/domain/post/post/entity/Post.java
@@ -42,8 +42,9 @@ public class Post extends BaseEntity {
 
     private boolean isPaid;
     private boolean isNotPublished;
-    private Integer countByMember;
-    private Integer viewCount;
+    private Long countByMember;
+    private Long viewCount;
+
 
     @OneToMany(mappedBy = "post", cascade = CascadeType.REMOVE)
     private List<Comment> commentList;
